==> 01_train_test_split.ipynb <==
import pandas as pd
import os
from sklearn.model_selection import train_test_split
import numpy as np

os.chdir('../../Data')

df = pd.read_csv('final_data.csv',header=0)

rename_dict = {col : '_'.join(col.split(' ')) for col in df.columns}

rename_dict

df.rename(rename_dict,copy=False,inplace=True,axis=1)

companies = list(df['Symbol'].unique())

cols_to_shift = [col for col in df.columns if col not in ['Date','Symbol','Series','Prev_Close','Open','Sector','Close']]

rename_dict2 = {col : col+'_1d' for col in cols_to_shift}

rename_dict2

development = pd.DataFrame()
for company in companies:
    print('development data for {}'.format(company))
    df_append = df[df['Symbol']==company]
    df_append['Date'] = pd.to_datetime(df['Date'])
    df_append.sort_values(by='Date',inplace=True)
    for index in ['nifty','auto','bank','fmcg','it','media','metal','pharma','realty','psu','500']:
        for i in range(1,6):
            df_append['return_{j}d_{index}'.format(j=i,index=index)] = (df_append['Close_{}'.format(index)].shift(i)-df_append['Open_{}'.format(index)].shift(i))/df_append['Open_{}'.format(index)].shift(i)*100
        df_append['Voltality_{}'.format(index)] = df_append[['return_{i}d_{index}'.format(i=i,index=index) for i in range(1,6)]].apply(lambda x:np.std(x),axis=1)
    for i in range(1,6):
        df_append['return_{j}d'.format(j=i)] = (df_append['Close'].shift(i)-df_append['Open'].shift(i))/df_append['Open'].shift(i)*100
    df_append['Voltality'] = df_append[['return_1d','return_2d','return_3d','return_4d','return_5d']].apply(lambda x:np.std(x),axis=1)
    df_append[['Open_1d']+cols_to_shift] = df_append[['Open']+cols_to_shift].shift(1)
    df_append.rename(rename_dict2,copy=False,inplace=True,axis=1)
    development = development.append(df_append.iloc[5:])

development.shape

development.columns

development['lift_ind'] = development[['Close','Open']].apply(lambda x: 1 if x[0] > x[1] else 0,axis=1)

development['lift_ind']

null_cols = []
for col in development.columns:
    if development[col].isnull().any():
        print(col)
        null_cols.append(col)

development.drop(null_cols,axis=1,inplace=True)

for col in development.columns:
    if development[col].isnull().any():
        print(col)

development.set_index(['Symbol','Date'],inplace=True)

open_price = development['Open']
close_price = development['Close']

development.drop(['Open','Close'],axis=1,inplace=True)

development.reset_index(inplace=True)

training = development.loc[development['Date']<'2019-01-01']
testing =  development.loc[development['Date']>='2019-01-01']

os.chdir('../')

training.shape

os.getcwd()

os.chdir('Probability Prediction/Input_Data')

open_price.to_csv('open_price.csv')
close_price.to_csv('close_price.csv')

training.to_csv('train.csv',index=False)

testing.to_csv('validation.csv',index=False)
