==> 07_reason_codes.ipynb <==
import pandas as pd
import numpy as np
import os
import pickle
from treeinterpreter import treeinterpreter as ti
from scipy.special import cbrt
import shap

from openpyxl import load_workbook,Workbook

# Getting data

os.chdir('../pickles')

numerical = pickle.load(open('numerical.pickle','rb'))
categorical = pickle.load(open('categorical.pickle','rb'))
conversion_dict = pickle.load(open('conversion_dict.pickle','rb'))
imputation_cols = pickle.load(open('imputation_cols.pickle','rb'))
model_columns = pickle.load(open('model_columns','rb'))
lr_model_columns = pickle.load(open('lr_model_columns.pickle','rb'))
order = pickle.load(open('order.pickle','rb'))

os.chdir('../Statistics/')

df_val = pd.read_excel('future_predictions.xlsx',sheet_name='data',header=0)

df_val.columns

os.chdir('../Input_Data/')

open_prices = pd.read_csv('open_price.csv',header=None)
close_prices = pd.read_csv('close_price.csv',header=None)

open_prices.columns = ['Symbol','Date','Open_Price']
close_prices.columns = ['Symbol','Date','Close_Price']

open_prices.set_index(['Symbol','Date'],inplace=True)
close_prices.set_index(['Symbol','Date'],inplace=True)

df_val.set_index(['Symbol','Date'],inplace=True)

df_val = df_val.join(open_prices,how='left')

df_val = df_val.join(close_prices,how='left')

# Reason Codes for logistic regression model

os.chdir('../Models/')

LR = pickle.load(open('LR.pickle','rb'))
RF = pickle.load(open('RF.pickle','rb'))
GBM = pickle.load(open('GBM.pickle','rb'))

df_val.reset_index(inplace=True)

df_val.columns

def reason_code_lr(model,data,target='lift_ind'):
    df_reasons = pd.DataFrame()
    df_reasons['Symbol'] = data['Symbol']
    df_reasons['Date'] = data['Date']
    df_reasons['Open Price'] = data['Open_Price']
    df_reasons['Prev Close'] = data['Prev_Close_order5']*(10**5)
    df_reasons['Close Price'] = data['Close_Price']
    df_reasons['lift_ind'] = data['lift_ind']
    df_reasons['Probability'] = model.predict_proba(data[lr_model_columns])[:,1]
    df_reasons['log_odds'] = df_reasons['Probability'].apply(lambda x: np.log(x/(1-x)))
    df_reasons['bias'] = model.intercept_[0]
    coef_array = np.array(model.coef_[0])
    values = np.array(data[lr_model_columns])
    contributions = np.array([coef_array*values[i,:] for i in range(values.shape[0])])
    contributions = pd.DataFrame(contributions)
    contributions['mrc'] = contributions.apply(lambda x: ','.join([a+':'+str(b) for (a,b) in sorted(zip(lr_model_columns,x),key = lambda x: x[1])]),axis=1)
    df_reasons['mrc'] = contributions['mrc']
    
    return df_reasons

df_reasons = reason_code_lr(LR,df_val)

os.chdir('../Statistics/')

book = load_workbook('future_predictions.xlsx')
writer = pd.ExcelWriter('future_predictions.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

df_reasons.to_excel(writer, "lr_reason_codes", index=False)

writer.save()

# Random Forest Reason Codes

def reason_code_rf(model,data,target='lift_ind'):
    predictors = lr_model_columns
    df_reasons = pd.DataFrame()
    prediction, bias, contributions = ti.predict(model,np.array(data[lr_model_columns]))
    df_reasons['Symbol'] = data['Symbol']
    df_reasons['Date'] = data['Date']
    df_reasons['Open Price'] = data['Open_Price']
    df_reasons['Prev Close'] = data['Prev_Close_order5']*(10**5)
    df_reasons['Close Price'] = data['Close_Price']
    df_reasons['lift_ind'] = data['lift_ind']
    df_reasons['Probability'] = model.predict_proba(data[lr_model_columns])[:,1]
    df_reasons['bias'] = bias[:,1]
    contributions = pd.DataFrame(contributions[:,:,1])
    contributions['mrc'] = contributions.apply(lambda x: ','.join([a+':'+str(b) for (a,b) in sorted(zip(lr_model_columns,x),key = lambda x: x[1])]),axis=1)
    df_reasons['mrc'] = contributions['mrc']
    
    return df_reasons

df_reasons = reason_code_rf(RF,df_val)

book = load_workbook('future_predictions.xlsx')
writer = pd.ExcelWriter('future_predictions.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

df_reasons.to_excel(writer, "rf_reason_codes", index=False)

writer.save()

# reason code for gbm model 

def reason_code_gbm(model,data,target='lift_ind'):
    predictors = lr_model_columns
    df_reasons = pd.DataFrame()
    df_reasons['Symbol'] = data['Symbol']
    df_reasons['Date'] = data['Date']
    df_reasons['Open Price'] = data['Open_Price']
    df_reasons['Prev Close'] = data['Prev_Close_order5']*(10**5)
    df_reasons['Close Price'] = data['Close_Price']
    df_reasons['lift_ind'] = data['lift_ind']
    df_reasons['Probability'] = model.predict_proba(data[lr_model_columns])[:,1]
    explainer = shap.TreeExplainer(model)
    shap_values = explainer.shap_values(data[model_columns])
    df_reasons['bias'] = explainer.expected_value
    contributions = pd.DataFrame(shap_values)
    contributions['mrc'] = contributions.apply(lambda x: ','.join([a+':'+str(b) for (a,b) in sorted(zip(lr_model_columns,x),key = lambda x: x[1])]),axis=1)
    df_reasons['mrc'] = contributions['mrc']
    
    return df_reasons

df_reasons = reason_code_gbm(GBM,df_val)

df_reasons

book = load_workbook('future_predictions.xlsx')
writer = pd.ExcelWriter('future_predictions.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

df_reasons.to_excel(writer, "gbm_reason_codes", index=False)

writer.save()

# The End
